clean up project structure
